Continue to refactor things
Mostly prep for a serious refactor session down the road, but we did do some refactoring, and we're making small progress...

Also update TODO.txt.
