+++ views/static/test/autosize-test.php
@@ -1,12 +1,15 @@
 <?php
-    require_once(dirname(__FILE__)."/../init.php");
+
+use App\Models\BlankableWord;
+
+require_once(dirname(__FILE__)."/../init.php");
     $sections = load_home_sections($pdo, $_SESSION["ConferenceID"]);
 
-    $words = load_non_blankable_words($pdo);
+    $words = BlankableWord::loadAllBlankableWords($pdo);
 
-    $question = generate_fill_in_question("There was a boy called Eustace Clarence Scrubb, and he almost deserved it.", 1, $words);
-    //$question = generate_fill_in_question("There...almost deserved it.", 0.75, words);
-    //$question = generate_fill_in_question("\"My dear boy, what ever shall you do?\"", 0.5, words);
+    $question = BlankableWord::generateFillInQuestion("There was a boy called Eustace Clarence Scrubb, and he almost deserved it.", 1, $words);
+    //$question = BlankableWord::generateFillInQuestion("There...almost deserved it.", 0.75, words);
+    //$question = BlankableWord::generateFillInQuestion("\"My dear boy, what ever shall you do?\"", 0.5, words);
 ?>
 
 <?php include(dirname(__FILE__)."/../header.php"); ?>
